1、版本号修改为2.1.0 2、增加网关流水线

+++ kongx-serve/src/main/java/com/kongx/serve/service/flow/ServicePipelineService.java
@@ -2,8 +2,10 @@ package com.kongx.serve.service.flow;
 
 import com.kongx.common.core.entity.UserInfo;
 import com.kongx.serve.entity.flow.ServicePipeline;
+import com.kongx.serve.entity.system.SystemProfile;
 import com.kongx.serve.mapper.ServicePipelineMapper;
 import com.kongx.serve.service.IBaseService;
+import com.kongx.serve.service.system.SystemProfileService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -15,17 +17,18 @@ public class ServicePipelineService implements IBaseService<ServicePipeline, Int
 
     @Autowired
     private ServicePipelineMapper servicePipelineMapper;
+    @Autowired
+    private SystemProfileService systemProfileService;
 
-
-    @Override
-    public List<ServicePipeline> findAll(ServicePipeline entity) {
-        return this.servicePipelineMapper.findAll(entity);
+    public List<ServicePipeline> findAll(SystemProfile systemProfile, ServicePipeline entity) {
+        return this.servicePipelineMapper.findAll(systemProfile, entity);
     }
 
     @Override
     public void add(ServicePipeline entity, UserInfo userInfo) {
         entity.setCreator(userInfo.getName());
         entity.setCreate_at(new Date());
+        entity.setProfile(systemProfileService.getClientByName(userInfo.getUserId()).getProfile());
         this.servicePipelineMapper.insert(entity);
     }
 
@@ -38,11 +41,19 @@ public class ServicePipelineService implements IBaseService<ServicePipeline, Int
 
     @Override
     public void remove(int pk) {
-        this.servicePipelineMapper.deleteById(pk);
+        ServicePipeline servicePipeline = this.servicePipelineMapper.findById(pk);
+        if (servicePipeline.getNodeList().isEmpty()) {
+
+            this.servicePipelineMapper.deleteById(pk);
+        } else {
+            throw new RuntimeException("请删除相关配置后再执行本操作");
+
+        }
     }
 
     @Override
     public ServicePipeline findById(Integer id) {
-        return this.servicePipelineMapper.findById(id);
+        ServicePipeline servicePipeline = this.servicePipelineMapper.findById(id);
+        return servicePipeline;
     }
 }
